requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Visualization & Dashboard
streamlit>=1.29.0
matplotlib>=3.8.0
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
openpyxl>=3.1.0  # For Excel export
flask>=3.0.0  # For internal API

# SMS & Scheduling
twilio>=8.10.0  # For SMS/texting
APScheduler>=3.10.0  # For scheduling hourly jobs

